Add files via upload
This program is designed to efficiently manage and navigate a comprehensive dataset of hotels, encompassing a variety of attributes such as name, star rating, location, among others. It facilitates the organisation and retrieval of hotel information based on these diverse features, thereby streamlining the process of hotel management and discovery.